% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_api.R
\name{AccountAPI}
\alias{AccountAPI}
\alias{list_accounts}
\alias{list_balances}
\alias{list_positions}
\title{Functions for the Accounts API}
\usage{
list_accounts()

list_balances(aid = NULL, asDT = FALSE)

list_positions(aid = NULL, postype = c("EQ", "ALL", "OPTN", "INDX", "MF",
  "FI"), asDT = FALSE)
}
\arguments{
\item{aid}{An accountId required to look up information specific to one or more accounts.}

\item{asDT}{TBD}

\item{postype}{The underlying asset type to retreive}
}
\description{
These functions wrap calls to the Accounts API. See details for additional information.
}
\section{Functions}{
\itemize{
\item \code{list_accounts}: list all accounts for the authenticated user

\item \code{list_balances}: Retrieve balance information specific to one or more accounts for the authenticated user

\item \code{list_positions}: List account positions
}}

