% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_session.R
\docType{data}
\name{etrade_session}
\alias{etrade_session}
\alias{ETRADE}
\title{Etrade Session Management}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ETRADE
}
\description{
Functions to manage the active session. Including functions to begin or renew a stale session
}
\keyword{datasets}
