% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etrade_api.R
\name{internal_api}
\alias{internal_api}
\alias{etrade_api}
\alias{etrade_url}
\alias{eval_api}
\title{Etrade API Call}
\usage{
etrade_api(module = c("accounts", "market", "order", "notification"),
  api = NULL, session = NULL, query = NULL, verb = httr::GET,
  norun = FALSE, url = NULL)

etrade_url(module = NULL, api, query = NULL)

eval_api(api_call)
}
\arguments{
\item{module}{The etrade API module to reach (one of: "accounts", "market", "order", "notification")}

\item{api}{The API endpoint to reach within the module}

\item{session}{An internal session object}

\item{query}{list of other arguments for the body of the call}

\item{verb}{A verb function from package httr (GET (default), POST, etc)}

\item{norun}{boolean (default = FALSE). If true, returns unevaluated api call with
environment captured, that can be executed later with function \code{eval_api}}

\item{url}{send in entire url string instead of using module and api}

\item{api_call}{unevaluted api call retuned by \code{etrade_api(..., norrun = TRUE)}}
}
\description{
A function to assist with lower-level etrade API calls.
}
\section{Functions}{
\itemize{
\item \code{etrade_api}: the function that creates an etrade api call

\item \code{etrade_url}: A helper function to build Etrade API call

\item \code{eval_api}: A helper function that evaluates an etrade API call
}}

