% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{lookup_functions}
\alias{lookup_functions}
\alias{lu_my_tickers}
\alias{lu_portfolio_stat}
\alias{lu_performance_pct}
\title{Lookup Etrade Functions}
\usage{
lu_my_tickers(asDT = FALSE)

lu_portfolio_stat()

lu_performance_pct()
}
\arguments{
\item{asDT}{Boolean. Default FALSE will return a character vector. TRUE will return
a one-column data.table}
}
\description{
Authenticated Functions to Lookup Summary Stats Regarding Current Assets.
}
\section{Functions}{
\itemize{
\item \code{lu_my_tickers}: get tickers as vector or column

\item \code{lu_portfolio_stat}: aggregates and calculates return and weights of all assets owned

\item \code{lu_performance_pct}: calculates and prints total net return
}}

