% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_api.R
\name{market_api}
\alias{market_api}
\alias{et_quote}
\alias{et_find_ticker}
\alias{et_quote_all}
\title{Get Up-to-Date Quote Information}
\usage{
et_quote(tickers = NULL, detailFlag = c("ALL", "FUNDAMENTAL", "INTRADAY",
  "WEEK_52"))

et_find_ticker(companyname = NULL, type = c("EQ", "MF"))

et_quote_all()
}
\arguments{
\item{tickers}{A character vector of tickers}

\item{detailFlag}{Optional. Can specify a subset of data provided by API.
Defaults to 'ALL' which returns all variables for given tickers}

\item{companyname}{A character vector specifying the company name to look the ticker up}

\item{type}{One of "EQ" for equity or "MF" for mutual fund}
}
\value{
A dataset of class data.table
}
\description{
This function will return a data.table containing quotes and additional
information, such as Fundementals or Intraday prices, for the provided
ticker or tickers.
}
\section{Functions}{
\itemize{
\item \code{et_quote}: Function to retrieve quotes and other information

\item \code{et_find_ticker}: Function search for ticker based on potential company name

\item \code{et_quote_all}: Function to pull all tickers in active account and retrieve quote and additional info for in real-time
}}

\examples{
# et_quote(c("BAC", "GOOG"))
}
