% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stamp.R
\name{print_stamp}
\alias{print_stamp}
\title{Utility to Print on Console}
\usage{
print_stamp(msg, sym = "#")
}
\arguments{
\item{msg}{Message to print}

\item{sym}{Symbol to use for readability}
}
\description{
Utility to Print on Console
}
