% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_functions.R
\name{resource_functions}
\alias{resource_functions}
\alias{rsrc_list_sic}
\alias{rsrc_list_etfs}
\alias{rsrc_list_pubco}
\title{External Resource Functions}
\usage{
rsrc_list_sic()

rsrc_list_etfs()

rsrc_list_pubco()
}
\description{
Functions that assist with interactive exploration and analysis. These functions are not
Etrade API client functions and pull resources externally from the web
}
\section{Functions}{
\itemize{
\item \code{rsrc_list_sic}: List all SIC codes

\item \code{rsrc_list_etfs}: List all ETFs

\item \code{rsrc_list_pubco}: List companies listed on NSDQ, NYSE, and AMEX exchanges.
}}

