% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_connect.R
\name{et_connect}
\alias{et_connect}
\alias{et_app}
\alias{et_request_token}
\alias{et_auth_url}
\alias{et_connect}
\alias{EtradeConnect}
\title{Internal Connection Functions}
\usage{
et_app()

et_request_token(app)

et_auth_url(app, rtoken)

et_connect()

EtradeConnect()
}
\description{
Internally used to facilitate back-end connection to Etrade
}
\section{Functions}{
\itemize{
\item \code{et_app}: returns etrade application saved in this package

\item \code{et_request_token}: function to ask for and retrieve request token from etrade

\item \code{et_auth_url}: Construct url for authorization call to Etrade

\item \code{et_connect}: Internal Streamlined Connection Without Shiny Application for Testing Purposes

\item \code{EtradeConnect}: Launches application to faciliate 2-factor authentication
}}

