% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_connect.R
\name{et_connect}
\alias{et_connect}
\alias{et_default_app}
\alias{et_connect}
\alias{stop_for_access}
\title{Internal Connection Functions}
\usage{
et_default_app()

et_connect()

stop_for_access()
}
\description{
Internally used to facilitate back-end connection to Etrade

#' @describeIn et_connect retrieves or initializes authentication to Etrade
#' @export
et_connect <- rg_connect
}
\section{Functions}{
\itemize{
\item \code{et_default_app}: sets system variables in order to use default app of this package

\item \code{et_connect}: retrieves or initializes authentication to Etrade

\item \code{stop_for_access}: returns ERROR if no access is found
}}

