% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{lookup_functions}
\alias{lookup_functions}
\alias{lu_all_tickers}
\alias{lu_portfolio_stat}
\alias{lu_performance_pct}
\alias{lu_all_SIC}
\alias{lu_etfs}
\alias{lu_all_pubco}
\title{Analysis Helpers}
\usage{
lu_all_tickers(asDT = FALSE)

lu_portfolio_stat()

lu_performance_pct()

lu_all_SIC()

lu_etfs()

lu_all_pubco()
}
\arguments{
\item{asDT}{Boolean. Default FALSE will return a character vector. TRUE will return
a one-column data.table}
}
\description{
Functions that assist with interactive exploration and analysis. Includes quick lookup
wrappers around main functions.
}
\details{
Includes account specific functions as well as general information lookup helpers.
}
\section{Functions}{
\itemize{
\item \code{lu_all_tickers}: get tickers as vector or column

\item \code{lu_portfolio_stat}: aggregates and calculates return and weights of all assets owned

\item \code{lu_performance_pct}: calculates and prints total net return

\item \code{lu_all_SIC}: List all SIC codes

\item \code{lu_etfs}: List all ETFs

\item \code{lu_all_pubco}: List companies listed on NSDQ, NYSE, and AMEX exchanges.
}}

